[EPMEDU-886]: Move Profile screen to separate module
