+++ feature/mainflow/moreflow/profile/src/main/java/com/epmedu/animeal/more/profile/ProfileScreen.kt
@@ -1,4 +1,4 @@
-package com.epmedu.animeal.more.profile.presentation
+package com.epmedu.animeal.more.profile
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
@@ -6,7 +6,7 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import com.epmedu.animeal.extensions.currentOrThrow
 import com.epmedu.animeal.foundation.bottombar.BottomBarVisibility
 import com.epmedu.animeal.foundation.bottombar.BottomBarVisibilityState.HIDDEN
-import com.epmedu.animeal.more.profile.presentation.viewmodel.ProfileViewModel
+import com.epmedu.animeal.more.profile.viewmodel.ProfileViewModel
 import com.epmedu.animeal.navigation.navigator.LocalNavigator
 
 @Composable
