+++ feature/mainflow/moreflow/profile/src/main/java/com/epmedu/animeal/more/profile/ProfileScreenEvent.kt
@@ -1,4 +1,4 @@
-package com.epmedu.animeal.more.profile.presentation
+package com.epmedu.animeal.more.profile
 
 internal sealed interface ProfileScreenEvent {
     object Edit : ProfileScreenEvent
