+++ feature/mainflow/moreflow/profile/src/main/java/com/epmedu/animeal/more/profile/ProfileScreenUI.kt
@@ -1,4 +1,4 @@
-package com.epmedu.animeal.more.profile.presentation
+package com.epmedu.animeal.more.profile
 
 import androidx.activity.compose.BackHandler
 import androidx.compose.runtime.Composable
@@ -10,8 +10,8 @@ import androidx.compose.ui.res.stringResource
 import com.epmedu.animeal.foundation.dialog.AnimealAlertDialog
 import com.epmedu.animeal.foundation.preview.AnimealPreview
 import com.epmedu.animeal.foundation.theme.AnimealTheme
-import com.epmedu.animeal.more.profile.presentation.ProfileScreenEvent.Discard
-import com.epmedu.animeal.more.profile.presentation.ui.ProfileContent
+import com.epmedu.animeal.more.profile.ProfileScreenEvent.Discard
+import com.epmedu.animeal.more.profile.ui.ProfileContent
 import com.epmedu.animeal.profile.presentation.ProfileInputFormEvent
 import com.epmedu.animeal.profile.presentation.viewmodel.ProfileState
 import com.epmedu.animeal.resources.R
