+++ feature/mainflow/moreflow/profile/src/main/java/com/epmedu/animeal/more/profile/ui/ProfileButton.kt
@@ -1,4 +1,4 @@
-package com.epmedu.animeal.more.profile.presentation.ui
+package com.epmedu.animeal.more.profile.ui
 
 import androidx.compose.foundation.layout.*
 import androidx.compose.material.Divider
