+++ settings.gradle.kts
@@ -47,11 +47,11 @@ include(":app")
 include(":feature:home")
 include(":feature:mainflow:moreflow:about")
 include(":feature:mainflow:moreflow:account")
-include(":feature:mainflow:moreflow:flow")
-include(":feature:mainflow:moreflow:more")
-include(":feature:mainflow:moreflow:help")
 include(":feature:mainflow:moreflow:donate")
-include(":feature:more")
+include(":feature:mainflow:moreflow:flow")
+include(":feature:mainflow:moreflow:help")
+include(":feature:mainflow:moreflow:more")
+include(":feature:mainflow:moreflow:profile")
 include(":feature:signupflow:entercode")
 include(":feature:signupflow:enterphone")
 include(":feature:signupflow:finishprofile")
